Maybe naïve is better

+++ src/torchCompactRadius/cppSrc/hashing.cpp
@@ -51,6 +51,6 @@ torch::Tensor computeHashIndices_t(torch::Tensor cellIndices, uint32_t hashMapLe
 }
 
 
-torch::Tensor computeHashIndices(torch::Tensor cellIndices, int64_t hashMapLength){
+torch::Tensor computeHashIndices(torch::Tensor cellIndices, int32_t hashMapLength){
     return computeHashIndices_t(cellIndices, (uint32_t) hashMapLength);
 }
